Alpha_Vantage_Pipeline/functions/parsev2/requirements.txt
functions-framework==3.*
google-cloud-secret-manager
google-cloud-storage
requests==2.28.1
pandas
google-cloud-bigquery

